Add Google Analytics tracking code to index.html
This commit adds the Google Analytics tracking code (gtag.js) to the index.html file. The code is added near the end of the file, before the closing body tag (</body>). The tracking ID used is "G-6ZW0PJ9LBN". The JavaScript code initializes the dataLayer array and configures gtag with the tracking ID.